adding initial random lookups
