+++ simulator/src/main/java/peersim/kademlia/das/DASProtocolValidator.java
@@ -1,7 +1,9 @@
 package peersim.kademlia.das;
 
 import java.math.BigInteger;
+import peersim.core.CommonState;
 import peersim.kademlia.Message;
+import peersim.kademlia.das.operations.ValidatorSamplingOperation;
 
 public class DASProtocolValidator extends DASProtocol {
 
@@ -47,8 +49,75 @@ public class DASProtocolValidator extends DASProtocol {
   }
 
   @Override
+  protected void handleInitNewBlock(Message m, int myPid) {
+    super.handleInitNewBlock(m, myPid);
+    logger.warning("Starting validator (rows and columns) sampling");
+    startRowsandColumnsSampling();
+    // logger.warning("Starting random sampling");
+    // startRandomSampling();
+  }
+
+  @Override
   protected void handleGetSampleResponse(Message m, int myPid) {
     logger.warning("Received sample validator node: do nothing");
+  }
+
+  /**
+   * Starts getting rows and columns, only for validators
+   *
+   * @param m initial message
+   * @param myPid protocol pid
+   */
+  protected void startRowsandColumnsSampling() {
+    logger.warning(
+        "Starting rows and columns fetch "
+            + rowWithHighestNumSamples()
+            + " "
+            + row[rowWithHighestNumSamples()]
+            + " "
+            + columnWithHighestNumSamples()
+            + " "
+            + column[columnWithHighestNumSamples()]);
+
+    // start 2 row 2 column Validator operation (1 row/column with the highest number of samples
+    // already downloaded and another random)
+    createValidatorSamplingOperation(
+        CommonState.r.nextInt(KademliaCommonConfigDas.BLOCK_DIM_SIZE) + 1, 0, time);
+    createValidatorSamplingOperation(
+        0, CommonState.r.nextInt(KademliaCommonConfigDas.BLOCK_DIM_SIZE) + 1, time);
+    createValidatorSamplingOperation(
+        CommonState.r.nextInt(KademliaCommonConfigDas.BLOCK_DIM_SIZE) + 1, 0, time);
+    createValidatorSamplingOperation(
+        0, CommonState.r.nextInt(KademliaCommonConfigDas.BLOCK_DIM_SIZE) + 1, time);
+  }
+
+  private void createValidatorSamplingOperation(int row, int column, long timestamp) {
+    ValidatorSamplingOperation op =
+        new ValidatorSamplingOperation(
+            this.getKademliaId(),
+            timestamp,
+            currentBlock,
+            searchTable,
+            row,
+            column,
+            this.isValidator,
+            this);
+    samplingOp.put(op.getId(), op);
+    logger.warning("Sampling operation started validator " + op.getId());
+
+    op.elaborateResponse(kv.getAll().toArray(new Sample[0]));
+    op.setAvailableRequests(KademliaCommonConfigDas.ALPHA);
+    /*while (!doSampling(op)) {
+      if (!op.increaseRadius(2)) {
+        logger.warning("Operation completed max increase");
+        samplingOp.remove(op.getId());
+        logger.warning("Sampling operation finished");
+        KademliaObserver.reportOperation(op);
+        break;
+      }
+      logger.warning("Increasing " + op.getRadius() + " " + op.getClass().getCanonicalName());
+    }*/
+    doSampling(op);
   }
 
   /*public void processEvent(Node myNode, int myPid, Object event) {
